Feature/refatoring staging views (#16)
* Update README.md

* feat: Adição da nova coluna de escolaridade na tabela Dim_Pessoas.
- Criação das novas tabelas fatos de beneficios, lógica de idade de dependentes e titulares
- Novas classificações como Conjuge e Filhos(a), tabela de valores do plano de sáude compativel com a ANS
- Calculo mensal atual.
- fixes #11
- fixes #12

* git commit -m "feat(bronze): consolida ecossistema de dados raw com dominios de RH e testes de qualidade

- refactor(beneficios): implementa granularidade por vida na Fato_Vidas_Beneficios com chave familiar.
- fix(core): corrige falha de temporalidade biológica nas idades utilizando relativedelta.
- feat(lms): cria Fato_Engajamento_LMS simulando funil de aprendizagem (evasão, notas e horas).
- feat(cursos): isola motor combinatório no modulo lms_course_generator e normaliza Dim_Cursos.
- feat(ats): desenvolve Fato_ATS_Funil com métricas agregadas de SLA e conversão de recrutamento.
- feat(absenteismo): implementa Fato_Absenteismo_Historico aplicando regras da CLT (fracionamento de férias e licenças orgânicas).
- test(dq): expande suite de Pytest com DuckDB para validação de SLAs, integridade temporal e regras de negócio.
- fix(test): corrige syntax error de parser SQL na inicialização da fixture de testes."
- fixes #12
- fixes #13
- fixes #14
- fixes #15

---------

+++ iceberg_rh_dbt/models/staging/stg_fato_requisicoes_vagas.sql
@@ -5,14 +5,14 @@ WITH source AS (
 ),
 renamed_and_casted AS (
     SELECT 
-        CAST(id_requisicao AS INT64) AS sk_requisicao,
+        CAST(id_requisicao AS INT) AS sk_requisicao,
         CAST(data_abertura AS DATE) AS data_abertura,
         CAST(data_fechamento AS DATE) AS data_fechamento,
         tipo_vaga,
-        CAST(id_area AS INT64) AS sk_area,
-        CAST(id_cargo AS INT64) AS sk_cargo,
+        CAST(id_area AS INT) AS sk_area,
+        CAST(id_cargo AS INT) AS sk_cargo,
         status,
-        CAST(id_contrato_preenchimento AS INT64) AS sk_contrato_preechimento
+        CAST(id_contrato_preenchimento AS INT) AS sk_contrato_preechimento
     FROM source
 )
 SELECT *
